build: /W4 clean for all first-party targets
Common, Flasher, Logger, VolvoDiag, VolvoFlasher and VolvoLogger now
compile at /W4 with zero warnings: narrowing fixed with explicit casts
(UDSMessage/makePassThruMsg DataSize, toVector bytes, UDSRequest timeouts,
TransferExit CRC), shadowed names renamed (decideCtrlAction's parameter,
the pin window start, the YAML node loop), the unused makePassThruMsgs
removed, and the two unavoidable deprecations (codecvt-based string
conversion, getenv in the script runner) locally silenced with a stated
reason.

+++ Common/src/protocols/UDSMessage.cpp
@@ -20,7 +20,7 @@ PASSTHRU_MSG toPassThruMsg(uint32_t msgId, const uint8_t* Data, size_t DataSize,
     result.TxFlags = Flags;
     result.Timestamp = 0;
     result.ExtraDataIndex = 0;
-    result.DataSize = DataSize + sizeof(msgId);
+    result.DataSize = static_cast<unsigned long>(DataSize + sizeof(msgId));
     result.Data[0] = (msgId >> 24) & 0xFF;
     result.Data[1] = (msgId >> 16) & 0xFF;
     result.Data[2] = (msgId >> 8) & 0xFF;
