+++ Logger/src/LogParameter.cpp
@@ -5,19 +5,19 @@ namespace {
 		if (isSigned) {
 			switch (size) {
 			case 1:
-				return static_cast<int8_t>(value);
+				return static_cast<float>(static_cast<int8_t>(value));
 			case 2:
-				return static_cast<int16_t>(value);
+				return static_cast<float>(static_cast<int16_t>(value));
 			case 3:
 				// A plain int32_t cast would read a raw 0x800000 as +8388608: the sign
 				// lives in bit 23 and must be extended from there.
-				return static_cast<int32_t>((value & 0x800000) ? (value | 0xFF000000) : value);
+				return static_cast<float>(static_cast<int32_t>((value & 0x800000) ? (value | 0xFF000000) : value));
 			default:
-				return static_cast<int32_t>(value);
+				return static_cast<float>(static_cast<int32_t>(value));
 			}
         }
 		else {
-			return value;
+			return static_cast<float>(value);
 		}
 	}
 } // namespace
